Make Spectrum type not Copy, adapt code accordingly
The Spectrum struct has 401 64 bit floats. This is over 3 kiB of data.
This is probably better handled without being a Copy type. Copy types
are treated differently by the compiler, and are essentially always
passed by value. For such a large type, working with references,
and explicitly cloning when you need more instances is likely better.
However, there are currently no benchmarks to prove this.